Put effects into a separate file
