+++ rollup.config.js
@@ -1,7 +1,7 @@
 import { terser } from 'rollup-plugin-terser';
 
 export default {
-  input: './src/custom-element.js',
+  input: './src/index.js',
   output: {
     file: './dist/custom-element.min.js',
     format: 'es',
