+++ src/index.js
@@ -1,0 +1,2 @@
+export { generateClass, define } from './custom-element';
+export { dispatchEventEffect } from './effects';
